+++ spec/controllers/api/v2/account_lists/donations_controller_spec.rb
@@ -11,18 +11,22 @@ describe Api::V2::AccountLists::DonationsController, type: :controller do
   let!(:donations) do
     create_list(:donation, 2, donor_account: donor_account,
                               designation_account: designation_account,
-                              amount: 10.00)
+                              amount: 10.00,
+                              donation_date: Date.new(2017, 2, 22))
   end
   let(:donation) { donations.first }
   let(:id) { donation.uuid }
 
   before do
+    donation.update(donation_date: Date.new(2017, 2, 20))
     account_list.designation_accounts << designation_account
     contact.donor_accounts << donor_account
   end
 
   let(:resource) { donation }
   let(:parent_param) { { account_list_id: account_list_id } }
+  let(:filter_param) { { donation_date: '2017-02-21..2017-02-23' } }
+  let(:sorting_param) { :donation_date }
   let(:correct_attributes) { attributes_for(:donation) }
   let(:incorrect_attributes) { { donation_date: nil } }
   let(:unpermitted_attributes) { nil }